// apps/web/src/common/types.ts
export type Option = Record<string, unknown> & {
  label?: string;
  value: string | number | null;
};
// ...
export type OptionWithIcon = Option & { icon?: React.ReactNode };
export type DropdownValueSetter = (value: string | Option | OptionWithIcon) => void;
export type MentionOption = OptionWithIcon & {
  type: string;
  value: string;
  description?: string;
};